implement PR review recommendations
